Fixes some notation, update formalism
Tweaks the definition of Sat to allow predicates to refer to any
variables in the store. Technically these variables should be live,
but our typing restriction should handle this. We also probably need
a well-formedness constraint to show that the predicates are only
refer to integer valued variables, leaving that for later