Adds pretty-printing inferred types
The types are stated in terms of (uninterpreted predicates).
To make this work, I added support for a general annotation
functionality when printing the ast. To make the interface more
friendly I also allowed for general format strings in a new
(separately compiled) pretty printing module.

Also, some UTF output because I can.