Reintroduces the idea of post update types
When we do this, we don't bother trying to propagate result type
information for immutable parameters into the callees type environment.
THIS IS POTENTIALLY INCOMPLETE: if a function diverges unless an
integer argument satisfies P, we chould potentially propagate
that information back to the post-call type state. But without
this optimization we ask Z3 to infer a LOT of new predicates.

We can potentially eliminate the need for this optimization if/when
we run ownership inference _before_ refinement inference.