Adds refinement that refers to function entry values
This is accomplished with a kind of gross hack. We introduce
a dummy variable for every leaf in every input leaf type, and
then equate the leaf types with these dummy variables. When constraining
the output type, we make the corresponding dummy variable one of the
arguments to the output refinement.

At calls, we introduce another dummy variable for each input leaf type,
equate it with the path for the leaf type (if applicable) and add
the dummy binding to the environment. This dummy variable is then
used as "pre-value" argument to the output parameter.