Propagate post call info to const vars
This makes our design decision to not dump the entire
environment into constraints sounds. Previously, we dumped
refinements on heap locations because the refinements on
those locations could potentially provide refinements on constant
locations in scope. However, any refinement on a heap location
must have been generated from variables in scope (because to refine
a heap location you have to write a variable into the heap).
If the heap refinement for location L contains refinements about
free variables X and Y _independent_ of the value of L, then
that refinement must be generated at some program point
where that information is derived from constant locations.
If that point is within a callee, by propagating out constant
location refinements, we can propagate this information to
contexts where it is necessary, and thus avoid dumping
the heap refinements into scope which SIGNIFICANTLY improves
inference speed.