New recursive refinements
This commit switches us to use the approach of Rondon in
their Type-Based Verification of Data Structures paper. Roughly,
for every recursive type, we unfold the type once. Then, in the
"outer type" we generate a "local refinement" which must hold for
every element in the unfolded data structure. This refinement is applied
to the corresponding position in the unfolded, "inner type". Further,
for each position in the inner type, we generate an "inductive refinement",
which refines just the current value in the "inner type" and may refer to
dependent tuple positions in the "outer type". In other words, the inductive
refinement refines positions in the inner type w.r.t dependent tuple values in
the outer position. When this unfolded recursive type is yet again unfolded, we
copy the current refinement in the inner position, and then apply a new instance
of the inductive refinement which refines the value w.r.t the "middle type" (which
was the inner type before unfolding) positions.