Fixes nullity tracking
This fixes how we track nullity, while simultaneously
paying down a LOT of technical debt we had accumulated.

Roughly, we do a better job of tracking when pointer contents
are definitely non-null and propagate this information
down into constraint generation. We further have nullity stacks
which record the dependencies between nullity flags. Basically, if
we have a value nested under n references of unknown nullity,
then we constrain the nth nullity flag to be true only when all
n-1 parent nullity flags are true. Otherwise, the refinements in
outer tuple positions in the n-1 references will be given no refinements,
and any inductive refinements under the nth reference will be
impossible to prove. This failure was preventing us from verifying
the sorted list.