Refactors out ownership inference
Ownership inference now runs in a completely different
module, and no longer depends on the inference module.
Ownership inference checks that the program is ownership
well-formed, and records at key program points whether
a type is top or not. Notice we do NOT generate more
precise information about the exact ownership at program points
because it is actually irrelevant for the purposes of refinemnt
inference; that module just needs to know where it can discard
refinement information. The ownership inference/solver records
this information at:
* type (re)generation:
  * folds (let mkref, assign)
  * joins (if/ifnull)
  * null types (ifnull, let null)
  * mkarray (although there should be NO nested ownership)
  * method entry/exit types
* type splitting:
  * assignments, let bindings, etc.
  * calls