Add zero ownership reads; remove ownership inference
This commit adapts the refinement inference to use the
ownership information generated by the ownership inference
tool. Accordingly, it removes all of the ownership inference
infrastructure from refinementtypes and inference,
and switches to simple live/not live flags. These flags
indiciate whether the reference has reliable refinement
information, i.e. if there exists a mutable aliased pointer
for that reference. The inference engine now checks that the
refinements for refinements with not live ownership are always
set to top. As a nice side effect, this SIGNIFICANTLY simplifies
a large portion of the code base, in particular the nightmarish
apply_matrix. The encodings are also generated more cleanly as well.