Adds unit
The only use for evar was a return, but any evar in a tail
position are now lifted to a return. So it suffices to evaluate
any side effects, and then introduce an explicit "and then nothing"
form, aka unit.

Note that this does NOT introduce a unit type into the language, unit in
the surface syntax is still interpreted to be 0 in tail positions.